New: `sc-api row().scrollTo()` method to move the table's start point to a specific row
Dev: Documentation for scroller options and API methods. Note that some
methods have been left undocumented as I'm not sure of their value in
the public interface - I've not come across a usecase for them yet and
the 1.10 Scroller methods are only now being documented, so they may be
dropped in a future release. Likewise `scroller().measure()` is now
`scroller.measure()` similar to other methods in Responsive etc for
consistency.