Fix: improve height measurement
This improves the height measurement when deriving from a string value.
Previously it would work correctly only with pixels, now it also works
with units 'vh' and 'rem'.

There is partial support for 'em' - it works as long as the parent
element of the scroller has the same font size as the body element