`mousedown` flag set to true but never reset
When using scroller within a searchpane, there's a label displayed at the bottom of the pane (completely at the bottom, not following the scroller like it should be in standard table). 

The annoying thing is that this label remains displayed even when stopping scrolling and displays a number which represents nothing to the user. 

This change at least stops it from being updated continuously, which is completely useless as the label is displayed at the complete bottom of the search pane. IMO there should also be an option within the scroller extension to disable this label completely, but that will be for another time :)

*I saw there's a mousedown flag reset in the `FixedColumn` extension... but not in this one.